security(s3): fix user profile images
